chore: separate to read and write contract services
