+++ dapp/src/components/Commit.astro
@@ -42,7 +42,7 @@ import PrimaryButton from "./utils/PrimaryButton.astro";
 </div>
 
 <script>
-  import { commitHash } from "../service/ProjectService";
+  import { commitHash } from "../service/WriteContractService";
   import { loadProjectInfo } from "../service/StateService";
   import { loadedPublicKey } from "./stellar-wallets-kit";
