+++ dapp/src/components/GetCommit.astro
@@ -25,7 +25,7 @@ import PrimaryButton from "./utils/PrimaryButton.astro";
 </div>
 
 <script>
-  import { getProjectHash } from "../service/ProjectService";
+  import { getProjectHash } from "../service/ReadContractService";
 
   document.addEventListener("astro:page-load", () => {
     const wrap_loader = document.getElementById(
