+++ dapp/src/components/RegisterProject.astro
@@ -64,7 +64,7 @@ import ModalInput from "./utils/ModalInput.astro";
 </div>
 
 <script>
-  import { registerProject } from "../service/ProjectService";
+  import { registerProject } from "../service/WriteContractService";
   import { navigate } from "astro:transitions/client";
   const SOROBAN_DOMAIN_CONTRACT_ID = `${import.meta.env.PUBLIC_SOROBAN_DOMAIN_CONTRACT_ID}`;
