+++ dapp/src/components/SetProject.astro
@@ -43,7 +43,7 @@ const TANSU_CONTRACT_ID = import.meta.env.PUBLIC_TANSU_CONTRACT_ID;
 </div>
 
 <script>
-  import { getProject } from "../service/ProjectService";
+  import { getProject } from "../service/ReadContractService";
   import { setProjectId, setProject, setProjectRepoInfo, setConfigData, refreshLocalStorage } from "../service/StateService";
   import { navigate } from "astro:transitions/client";
   import { getAuthorRepo } from "../service/utils";
